Updated the projectile spell system to be more general and have its own cast spell function

+++ include/constants.h
@@ -51,10 +51,10 @@ constexpr float kTexCoordRight = 1.0f;
 constexpr int kProjectileVertices = 6;
 constexpr int kFireballWidth = 50;
 constexpr int kFireballHeight = 50;
-constexpr int kFireballSpeed = 350.0f;
+constexpr float kFireballSpeed = 350.0f;
 constexpr int kFrostboltWidth = 50;
 constexpr int kFrostboltHeight = 50;
-constexpr int kFrostboltSpeed = 250.0f;
+constexpr float kFrostboltSpeed = 250.0f;
 
 // Entity constants
 constexpr int kNumEntities = kNumEnemies + 1;
